refactor

+++ orchesrator.py
@@ -1,7 +1,7 @@
 import boto3
 from organization import Organization 
 from session_manager import session_generator
-from actions.iam_role import delete_iam_role
+from actions.iam_role import delete_iam_role, iam_role_exists
 
 REGIONS = ['us-east-1', 'eu-west-1', 'eu-central-1', 'ca-central-1', 'ap-southeast-1']
 REGIONS = ['eu-west-1']
@@ -15,10 +15,13 @@ def main():
         print("action")
         # clean_kms_keys(target_session)
         # delete_stacks_with_name(target_session, "StackSet-core-deployments-servicecatalog-role")
-        approve = input("Delete IAM role? (y/n): ")
-        if approve == "y":
-            print("Deleting IAM role")
-            delete_iam_role(target_session, "mfs-devops-ci-role")    
+        if iam_role_exists(target_session, "mfs-devops-ci-role"):
+            approve = input("Delete IAM role? (y/n): ")
+            if approve == "y":
+                print("Deleting IAM role")
+                delete_iam_role(target_session, "mfs-devops-ci-role")    
+        else:
+            print("IAM role does not exist")
 
 if __name__ == '__main__':
     main()
